Refactor Phoenix contexts
