+++ .iex.exs
@@ -1,6 +1,9 @@
 alias Meadow.Repo
 alias Meadow.{Accounts, Ingest}
-alias Meadow.Acccounts.User
-alias Meadow.Ingest.{Bucket, IngestJob, Project}
+alias Meadow.Acccounts.Users
+alias Meadow.Acccounts.Users.User
+alias Meadow.Ingest.{IngestJobs, Projects}
+alias Meadow.Ingest.Projects.{Bucket, Project}
+alias Meadow.Ingest.IngestJobs.{IngestJob, IngestRow, InventoryValidator}
 
 import Ecto.Query
